requirements.txt
streamlit
vecs
pydantic
numpy
requests
torch
transformers
PyPDF2
nltk
PyYAML
python-docx
youtube-transcript-api